:heavy_plus_sign: manage bracket in execution
*
