+++ src/exec/execute.c
@@ -6,32 +6,11 @@
 /*   By: foctavia <foctavia@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/08 17:35:56 by owalsh            #+#    #+#             */
-/*   Updated: 2022/09/16 19:54:50 by foctavia         ###   ########.fr       */
+/*   Updated: 2022/09/19 14:00:25 by foctavia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-int	ms_bracket(t_cmdlst **cmds, char **env)
-{
-	int	res;
-
-	res = EXIT_SUCCESS;
-	while (*cmds && (*cmds)->type != CLOSE_BRK)
-	{
-		if ((*cmds)->type == WORD)
-			res = exec_cmd(cmds, env);
-		if ((*cmds)->next && (*cmds)->next->next && (*cmds)->next->type == OPERAND && !res)
-			*cmds = (*cmds)->next->next;
-		else if ((*cmds)->next && (*cmds)->next->next && (*cmds)->next->type == OPEROR && res)
-			*cmds = (*cmds)->next->next;
-		else
-			break ;
-	}
-	while (*cmds && (*cmds)->type != CLOSE_BRK)
-		*cmds = (*cmds)->next;
-	return (res);
-}
 
 int	ms_execute(t_cmdlst **cmds, char **env)
 {
@@ -46,8 +25,8 @@ int	ms_execute(t_cmdlst **cmds, char **env)
 		res = exec_cmd(cmds, env);
 	if ((*cmds)->type == OPEN_BRK)
 		res = ms_bracket(&(*cmds)->next, env);
-	// if ((*cmds)->next && (*cmds)->next->type == CLOSE_BRK)
-	// 	return (res);
+	else if ((*cmds)->next && (*cmds)->next->type == OPEN_BRK)
+		res = ms_bracket(&(*cmds)->next->next, env);
 	if ((*cmds)->next && (*cmds)->next->type == OPERAND && !res)
 		res = ms_execute(&(*cmds)->next, env);
 	else if ((*cmds)->next && (*cmds)->next->type == OPEROR && res)
