+++ src/minishell.c
@@ -6,7 +6,7 @@
 /*   By: foctavia <foctavia@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/22 14:02:04 by owalsh            #+#    #+#             */
-/*   Updated: 2022/09/16 17:30:24 by foctavia         ###   ########.fr       */
+/*   Updated: 2022/09/19 11:36:33 by foctavia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,6 +34,7 @@ int	main(int argc, char **argv, char **env)
 			&& !ms_expanser(&data.tokens, res))
 		{
 			ms_parser(data.tokens, &data.cmds);
+			// display_cmds();
 			res = ms_execute(&data.cmds, data.shell.env);
 			res = ms_wait(&data.cmds, res);
 		}
